datasets merged
Some usersnames refer to an entity, such as the European Bank or the LOL's (videogame) website, not a person. And also, in the case of the vice-president of USA, instead of her own twitter account, the governmental vp account was retrieved and I guess that pattern is preserved for some others.